Refactoring the 'portfolio_history.py', 'portfolio_investment.py' and 'short_summary_tab.py' files
Refactoring the 'portfolio_history.py', 'portfolio_investment.py' and 'short_summary_tab.py' files
